Modularization of components was carried out, data separation, from the path components/CourseGrid.js.
