+++ components/grids/ComplementaryGrid.js
@@ -1,0 +1,10 @@
+"use client";
+
+import { useMemo } from "react";
+import { complementary } from "@/data/complementary";
+import GridContainer from "./GridContainer";
+
+export function ComplementaryGrid() {
+  const memoizedCourses = useMemo(() => complementary, []);
+  return <GridContainer items={memoizedCourses} />;
+}
