+++ components/grids/DocumentGrid.js
@@ -1,0 +1,10 @@
+"use client";
+
+import { useMemo } from "react";
+import { documents } from "@/data/documents";
+import GridContainer from "./GridContainer";
+
+export function DocumentGrid() {
+  const memoizedCourses = useMemo(() => documents, []);
+  return <GridContainer items={memoizedCourses} />;
+}
