+++ data/complementary.js
@@ -1,0 +1,10 @@
+export const complementary = [
+  {
+    title: "Automatización de procesos para la eficiencia organizacional",
+    link: "/docs/sena/fc_docs/automatizacion-de-procesos-para-la-eficiencia-organizacional",
+    images: {
+      light: "/images/index-images/documentacion/008.svg",
+      dark: "/images/index-images/documentacion/008.svg",
+    },
+  },
+];
