+++ data/courses.js
@@ -1,0 +1,35 @@
+// data/courses.ts
+export const courses = [
+  {
+    title: "Curso Git",
+    link: "./docs/cursos/git_curso",
+    images: {
+      light: "/images/index-images/cursos/git-logo.svg",
+      dark: "/images/index-images/cursos/git-logo-dark.svg",
+    },
+  },
+  {
+    title: "Curso Github",
+    link: "./docs/cursos/github_curso",
+    images: {
+      light: "/images/index-images/cursos/github-logo.svg",
+      dark: "/images/index-images/cursos/github-logo-dark.svg",
+    },
+  },
+  {
+    title: "Curso Conventional Commits",
+    link: "./docs/cursos/cc_curso",
+    images: {
+      light: "/images/index-images/cursos/conventional-commits-logo.svg",
+      dark: "/images/index-images/cursos/conventional-commits-logo-dark.svg",
+    },
+  },
+  {
+    title: "Shift Left Testing en Desarrollo Ágil",
+    link: "./docs/cursos/shift-left-testingen-desarrolloagil",
+    images: {
+      light: "/images/index-images/cursos/ShiftLeftTesting.svg",
+      dark: "/images/index-images/cursos/ShiftLeftTesting.svg",
+    },
+  },
+];
