+++ data/documents.js
@@ -1,0 +1,66 @@
+export const documents = [
+  {
+    title: "Sistemas operativos de red",
+    link: "./docs/sena/ads_docs/sistemas-operativos-de-red",
+    images: {
+      light: "/images/index-images/documentacion/001.svg",
+      dark: "/images/index-images/documentacion/001.svg",
+    },
+  },
+  {
+    title: "Redes y networking",
+    link: "./docs/sena/ads_docs/redes-y-networking",
+    images: {
+      light: "/images/index-images/documentacion/002.svg",
+      dark: "/images/index-images/documentacion/002.svg",
+    },
+  },
+  {
+    title: "Paradigma orientado a objetos",
+    link: "./docs/sena/ads_docs/paradigma-orientado-a-objetos",
+    images: {
+      light: "/images/index-images/documentacion/006.svg",
+      dark: "/images/index-images/documentacion/006.svg",
+    },
+  },
+  {
+    title: "Contenedores",
+    link: "./docs/sena/ads_docs/contenedores-documentos",
+    images: {
+      light: "/images/index-images/documentacion/004.svg",
+      dark: "/images/index-images/documentacion/004.svg",
+    },
+  },
+  {
+    title: "Virtualización",
+    link: "./docs/sena/ads_docs/virtualizacion",
+    images: {
+      light: "/images/index-images/documentacion/003.svg",
+      dark: "/images/index-images/documentacion/003.svg",
+    },
+  },
+  {
+    title: "Documentación y entrega",
+    link: "/docs/sena/actividades-de-documentacion-y-entrega-de-software-en-procesos-de-implantacion/1-introduccion",
+    images: {
+      light: "/images/index-images/documentacion/005.svg",
+      dark: "/images/index-images/documentacion/005.svg",
+    },
+  },
+  {
+    title: "Fundamentos de calidad del software",
+    link: "/docs/sena/ads_docs/fundamentos-de-la-calidad-del-software",
+    images: {
+      light: "/images/index-images/documentacion/007.svg",
+      dark: "/images/index-images/documentacion/007.svg",
+    },
+  },
+  {
+    title: "Aplicación de pruebas de software",
+    link: "/docs/sena/actividades-de-documentacion-y-entrega-de-software-en-procesos-de-implantacion/1-introduccion",
+    images: {
+      light: "/images/index-images/documentacion/008.svg",
+      dark: "/images/index-images/documentacion/008.svg",
+    },
+  },
+];
